agro/requirements.txt
fastapi
uvicorn
python-dotenv
pymysql
bcrypt
python-jose
